src/pyproject.toml
[project]
name = "econflux"
version = "0.1.0"
description = "Agentic financial intelligence system powered by Strands-Agents, yfinance, and Bedrock AgentCore."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=9.0.1",
    "strands-agents>=1.18.0",
    "bedrock-agentcore>=1.0.6",
    "strands-agents-tools>=0.2.16",
    "faker>=38.2.0",
    "boto3>=1.41.2",
    "botocore>=1.41.2",
    "dotenv>=0.9.9",
    "requests>=2.32.5",
    "pandas>=2.3.3",
]

[dependency-groups]
dev = [
    "bedrock-agentcore-starter-toolkit>=0.2.0",
]

